follow MSG generated code differences
Foo *getFoo() --> const Foo *getFoo() | Foo *getFooForUpdate()
bool getIsFoo() --> bool isFoo()
vector index type in MSG: unsigned int --> size_t
new vector functions:
  - insertFoo(Foo*)
  - insertFoo(size_t K, Foo*)
  - eraseFoo(size_t K)

+++ src/inet/networklayer/ipv6/IPv6ExtensionHeaders.cc
@@ -19,7 +19,7 @@
 
 namespace inet {
 
-void IPv6RoutingHeader::setAddressArraySize(unsigned int size)
+void IPv6RoutingHeader::setAddressArraySize(size_t size)
 {
     IPv6RoutingHeader_Base::setAddressArraySize(size);
     setByteLength(8 + 16 * size);
