+++ src/inet/networklayer/ipv6/IPv6ExtensionHeaders.h
@@ -18,7 +18,7 @@
 #ifndef __INET_IPV6EXTENSIONHEADERS_H
 #define __INET_IPV6EXTENSIONHEADERS_H
 
-#include "inet/networklayer/ipv6/IPv6ExtensionHeaders_m.h"
+#include "inet/networklayer/contract/ipv6/IPv6ExtensionHeaders_m.h"
 
 namespace inet {
 
@@ -30,7 +30,7 @@ class INET_API IPv6RoutingHeader : public IPv6RoutingHeader_Base
     IPv6RoutingHeader& operator=(const IPv6RoutingHeader& other) { IPv6RoutingHeader_Base::operator=(other); return *this; }
     virtual IPv6RoutingHeader *dup() const override { return new IPv6RoutingHeader(*this); }
     // ADD CODE HERE to redefine and implement pure virtual functions from IPv6RoutingHeader_Base
-    virtual void setAddressArraySize(unsigned int size) override;
+    virtual void setAddressArraySize(size_t size) override;
 };
 
 } // namespace inet
